bit of docs

+++ source/html/html.c
@@ -81,6 +81,19 @@ void Html_strset(const char* pDst, char ch, uint8_t maxlen)
 
 // === PUBLISHED FUNCTIONS ======================================================================================
 
+
+/*
+ * === Html_SetKeyValueInBuffer
+ * This functions sets value of given key in valueBuffer
+ * dictionary. The buffer is then used when copying HTML from
+ * its source to MTU buffer right before sending.
+ *
+ * Parameters:
+ *      key[in]      - key as which value is going to be changed
+ *      value[in]    - value to be changed
+ * Returns:
+ *      N/A
+ */
 void Html_SetKeyValueInBuffer(char key, char* value)
 {
     Html_strset(valueBuffer[key - 'a'].buf, 0, MAXLEN);
@@ -90,7 +103,21 @@ void Html_SetKeyValueInBuffer(char key, char* value)
     return;
 }
 
-uint8_t Html_CopyHtmlToMtuBuffer(uint16_t offset)
+
+/*
+ * === Html_CopyHtmlToMtuBuffer
+ * Copies HTML code from its source on the flash to the MTU
+ * buffer. While copying, if it comes across SUBSTITUE_TOKEN,
+ * it injects the value from valueBuffer dictionary.
+ *
+ * Parameters:
+ *      offset[in]   - used to offset to start of HTML code. Usable
+ *                     when splitting HTML code to more messages.
+ * Returns:
+ *      uint16_t     - number of copied characters
+ *
+ */
+uint16_t Html_CopyHtmlToMtuBuffer(uint16_t offset)
 {
     char *pHtml = (char*)(HTML_MEM_START + offset);
     char *pMtu  = (char*)MTU_BUF_MEM_START;
@@ -98,6 +125,10 @@ uint8_t Html_CopyHtmlToMtuBuffer(uint16_t offset)
 
     for ( i = 0; i < MTU_SIZE - 1; i++ )
     {
+        //
+        // Source HTML ended, remove all remaining characters from MTU buffer
+        // and cease copying.
+        //
         if ( *pHtml == '\0' )
         {
             Html_strset(pMtu, 0, MTU_SIZE - i);
@@ -118,6 +149,10 @@ uint8_t Html_CopyHtmlToMtuBuffer(uint16_t offset)
         }
     }
 
+    //
+    // Make sure that the last character from MTU buffer is '\0'
+    // so that other functions can print the buffer successfully.
+    //
     *pMtu = 0;
 
     return i;
